feat: Add selectedCategories and handleCategoryChange to GuideEditor
The changes in this commit add the `selectedCategories` prop and `handleCategoryChange` function to the `GuideEditor` component. These additions allow for the selection and handling of categories in the guide editor.
